Extract post persistence into repository mixin
