+++ changelog.d/+post-repository-slice.changed.md
@@ -1,0 +1,1 @@
+Moved post and post revision persistence into a dedicated repository slice while preserving the forum repository facade.
