+++ pyproject.toml
@@ -92,7 +92,6 @@ ignore = [
 "__init__.py" = ["F401"]
 "tests/**/*.py" = ["S101"]
 "scripts/**/*.py" = ["T20"]
-"src/elbysodic/db/repository.py" = ["S608"]
 "src/elbysodic/db/repositories/facets.py" = ["S608"]
 "src/elbysodic/db/repositories/materials.py" = ["S608"]
 "src/elbysodic/db/repositories/plot_hooks.py" = ["S608"]
